Move coda back to main package
