+++ main.go
@@ -2,16 +2,16 @@ package main
 
 import (
 	"bufio"
+	"errors"
 	"fmt"
-	"kotalbert/in-memory-notepad/notepad"
 	"os"
 	"strings"
 )
 
 func main() {
-	npd := notepad.NewNotepad()
+	npd := NewNotepad()
 	for {
-		fmt.Print("Enter a command and data:")
+		fmt.Print("Enter a command and data: ")
 		userInput := getInput()
 		command, data := parseInput(userInput)
 
@@ -32,7 +32,7 @@ func main() {
 			fmt.Print("[Info] Bye!\n")
 			os.Exit(0)
 		default:
-			fmt.Print(command, data)
+			fmt.Print("[Error] Unknown command\n")
 		}
 	}
 }
@@ -47,4 +47,53 @@ func getInput() string {
 	scanner := bufio.NewScanner(os.Stdin)
 	scanner.Scan()
 	return scanner.Text()
+}
+
+type Note struct {
+	Id   int
+	Text string
+}
+
+func (n Note) ToString() string {
+	return fmt.Sprintf("[Info] %d: %s", n.Id, n.Text)
+}
+
+const maxNoteNumber = 5
+
+type Notepad struct {
+	notes []Note
+}
+
+func NewNotepad() *Notepad {
+	return &Notepad{}
+}
+
+func (n *Notepad) CreateNote(text string) error {
+	if len(n.notes) >= maxNoteNumber {
+		return errors.New("the notepad is full")
+	}
+	n.notes = append(n.notes, Note{Id: len(n.notes) + 1, Text: text})
+	return nil
+}
+
+// ToString returns a string representation of the notepad
+//
+//	That is, a numbered list of all notes
+func (n *Notepad) ToString() string {
+	var sb strings.Builder
+
+	for _, note := range n.notes {
+		sb.WriteString(note.ToString())
+		sb.WriteString("\n")
+	}
+	return sb.String()
+}
+
+// ListNotes prints all notes to the console
+func (n *Notepad) ListNotes() {
+	fmt.Print(n.ToString())
+}
+
+func (n *Notepad) ClearNotes() {
+	n.notes = []Note{}
 }
