+++ notepad_test.go
@@ -1,4 +1,4 @@
-package notepad
+package main
 
 import (
 	"fmt"
@@ -14,7 +14,7 @@ func TestNotepadConstructor(t *testing.T) {
 
 func TestCreateNote(t *testing.T) {
 	notepad := NewNotepad()
-	notepad.CreateNote("test")
+	_ = notepad.CreateNote("test")
 	if len(notepad.notes) != 1 {
 		t.Error("Expected note to be created")
 	}
@@ -22,7 +22,7 @@ func TestCreateNote(t *testing.T) {
 
 func TestCreatedNoteShouldHaveValidData(t *testing.T) {
 	notepad := NewNotepad()
-	notepad.CreateNote("test")
+	_ = notepad.CreateNote("test")
 	if notepad.notes[0].Text != "test" {
 		t.Error("Expected note to have valid data")
 	}
@@ -30,8 +30,8 @@ func TestCreatedNoteShouldHaveValidData(t *testing.T) {
 
 func TestNotepad_ToString(t *testing.T) {
 	notepad := NewNotepad()
-	notepad.CreateNote("test1")
-	notepad.CreateNote("test2")
+	_ = notepad.CreateNote("test1")
+	_ = notepad.CreateNote("test2")
 
 	expected := fmt.Sprintf("[Info] 1: test1\n[Info] 2: test2\n")
